fix: remove bogus SYS_UTIMENSAT/SYS_FUTIMENS macOS syscall numbers
Root cause: macOS has no kernel syscall for utimensat (476=getattrlistat) or
futimens (477=proc_trace_log). These are libc wrappers over setattrlist.

Changes:
- Replace raw asm wrappers with dlsym-based libc passthrough
- Add REAL_FUTIMENS to reals.rs for dlsym resolution
- Add early-init guards to futimens_inception/utimensat_inception
- Move IT_UTIMENSAT/IT_FUTIMENS to __nointerpose (prevent recursion)
- Fix fcntl arg type (c_int -> i64) for macOS aarch64 ABI
- Simplify VFS prefix init (remove raw_realpath canonicalization)
- Fix clippy: empty-line-after-outer-attr, manual_strip

+++ crates/vrift-inception-layer/src/ipc.rs
@@ -49,7 +49,7 @@ pub(crate) unsafe fn raw_unix_connect(path: &str) -> c_int {
             let s = libc::socket(libc::AF_UNIX, libc::SOCK_STREAM, 0);
             if s >= 0 {
                 // BUG-007b: Use raw_fcntl via RawContext to avoid interposed fcntl
-                CTX.fcntl(s, libc::F_SETFD, libc::FD_CLOEXEC);
+                CTX.fcntl(s, libc::F_SETFD, libc::FD_CLOEXEC as i64);
             }
             s
         }
